fix(sync): restore toFloat() for totalVolumeKg in SqlDelightSyncRepository
GamificationStatsSyncDto.totalVolumeKg was changed from Int to Float
in SyncModels.kt but the SqlDelightSyncRepository call site was left
using toInt(), causing a compile error.

+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/data/repository/SqlDelightSyncRepository.kt
@@ -148,7 +148,7 @@ class SqlDelightSyncRepository(
                     clientId = row.id.toString(),
                     totalWorkouts = row.totalWorkouts.toInt(),
                     totalReps = row.totalReps.toInt(),
-                    totalVolumeKg = row.totalVolumeKg.toInt(),
+                    totalVolumeKg = row.totalVolumeKg.toFloat(),
                     longestStreak = row.longestStreak.toInt(),
                     currentStreak = row.currentStreak.toInt(),
                     updatedAt = row.updatedAt ?: row.lastUpdated
